Enable laser when checking size on Atomstack A5
Current code does not use gcodeStart/gcodeEnd when doing a check size.
For Atomstack A5, for the laser to start, the commands `M3 S0 F1000`
need to be sent at the beginning, e.g., in `gcodeStart`.

Since I'm not sure how to include properly gcodeStart in the checksize
gcode, I've naively hardcoded the necessary gcodeStart/gcodeEnd of
AtomStack A5.

Ideally, this PR needs to be changed to instead include
gcodeStart/gcodeEnd in the checkSize gcode.